making into reusable component
