+++ src/main/resources/i18n/messages.properties
@@ -361,7 +361,6 @@ ShareLayout.previous=Previous
 ShareLayout.next=Next
 ShareNoSamples.description=You have no samples selected to share, please go back to the project samples page and select some samples
 ShareNoSamples.link=Go to Project Samples Page
-ShareProject.label.id=id: {0}
 ShareSamples.title=Share Samples with Another Project
 ShareSamples.projects=Select a project to share samples with
 ShareSamples.ready=These samples are ready to copy
@@ -692,6 +691,9 @@ project.nav.exports=NCBI Exports
 project.nav.details=Recent Activity
 project.nav.settings=Settings
 project.nav.samples.import-metadata=Import Sample Metadata
+
+# Project select
+ProjectSelect.label.id=id: {0}
 
 # KEPT FOR LEGACY PURPOSES (still used on user_details.html and groups.html)
 project.table.collaborator.name=Name
